+++ src/CONFIG_TYPE.mjs
@@ -1,7 +1,9 @@
+export const CONFIG_TYPE_ARRAY = "array";
+
 export const CONFIG_TYPE_BOOLEAN = "boolean";
 
-export const CONFIG_TYPE_NONE = "none";
-
 export const CONFIG_TYPE_NUMBER = "number";
+
+export const CONFIG_TYPE_OBJECT = "object";
 
 export const CONFIG_TYPE_STRING = "string";