+++ src/ValueProvider/PREFIXES.mjs
@@ -1,5 +1,0 @@
-export const PREFIX_NO = "no-";
-
-export const PREFIXES = Object.freeze([
-    PREFIX_NO
-]);