+++ src/ValueProvider/SUFFIX.mjs
@@ -3,9 +3,3 @@ export const SUFFIX_FILE = "-file";
 export const SUFFIX_JSON = "-json";
 
 export const SUFFIX_JSON_FILE = `${SUFFIX_JSON}${SUFFIX_FILE}`;
-
-export const SUFFIXES = Object.freeze([
-    SUFFIX_FILE,
-    SUFFIX_JSON,
-    SUFFIX_JSON_FILE
-]);