+++ src/ValueProvider/ValueProvider.mjs
@@ -1,5 +1,5 @@
 /** @typedef {import("../FluxConfig.mjs").FluxConfig} FluxConfig */
 
 /**
- * @typedef {{getConfig: (key: string, flux_config: FluxConfig) => Promise<*>}} ValueProvider
+ * @typedef {{cast_types?: string[] | null, getConfig: (key: string, flux_config: FluxConfig) => Promise<*>}} ValueProvider
  */